Settings now uses a .ini file for all platforms
Addresses #19.

A side-effect of implementing this is that all platforms
now use a `.ini` file. Unix/Linux used to use a `.conf` file. I would
suggest to copy all the contents of the .conf file into a .ini file of
the same name. If not done, the setup process will run once again and
possibly overwrite the database and settings.